RFC2863_IF_MIB.pm:
The behavior of the builtin sort function is not defined if called in scalar context. So if you write a subroutine that directly returns the result of a sort operation, then you code will behave unpredictably if someone calls your subroutine in a scalar context. This Policy emits a violation if the return keyword is directly followed by the sort function. To safely return a sorted list of values from a subroutine, you should assign the sorted values to a temporary variable first. For example:

   sub frobulate {

       return sort @list;  # not ok!

       @sorted_list = sort @list;
       return @sort        # ok
   }

See also

http://perlcritic.com/pod/Perl/Critic/Policy/Subroutines/ProhibitReturnSort.html