XMLCompiler.pm: Non lexical loop values removed.
for/foreach loops always create new lexical variables for named iterators. In other words

    for $zed (...) {
        ...
    }

is equivalent to

    for my $zed (...) {
        ...
    }

This may not seem like a big deal until you see code like

    my $bicycle;
    for $bicycle (@things_attached_to_the_bike_rack) {
        if (
                $bicycle->is_red()
            and $bicycle->has_baseball_card_in_spokes()
            and $bicycle->has_bent_kickstand()
        ) {
            $bicycle->remove_lock();

            last;
        }
    }

    if ( $bicycle and $bicycle->is_unlocked() ) {
        ride_home($bicycle);
    }

which is not going to allow you to arrive in time for dinner with your family because the $bicycle outside the loop is different from the $bicycle inside the loop. You may have freed your bicycle, but you can't remember which one it was.

See also http://perlcritic.com/pod/Perl/Critic/Policy/Variables/RequireLexicalLoopIterators.html